+++ Binary Tree/query.h
@@ -1,6 +1,6 @@
 #include<iostream>
 #include "custom.h"
 
-Node* Search(Node* root, int val);
+Node* Search(Node* root, float val);
 Node* Minimum(Node* root);
 Node* Maximum(Node* root);
